requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

# ── HTTP client (for Ollama proxy calls) ──────────────────────────────────────
httpx==0.27.2

# ── Database ──────────────────────────────────────────────────────────────────
asyncpg==0.29.0

# ── Cache ─────────────────────────────────────────────────────────────────────
redis[asyncio]==5.0.8

# ── Config ────────────────────────────────────────────────────────────────────
pyyaml==6.0.2

# ── CLI (eval pipeline) ───────────────────────────────────────────────────────
typer==0.12.5
rich==13.8.1                    # pretty CLI output for eval scorecards

# ── ML / Evaluators ───────────────────────────────────────────────────────────

# PyTorch — Apple Silicon MPS build
# IMPORTANT: Do NOT install the CUDA build on Mac.
# The MPS build is the default for macOS on pip.
torch==2.4.1
torchvision==0.19.1             # required by some HuggingFace models

# HuggingFace
transformers==4.44.2
tokenizers==0.19.1
huggingface-hub==0.24.6
accelerate==0.34.2              # needed for device_map and MPS support

# ONNX Runtime (for optimized CPU inference)
onnxruntime==1.19.2
optimum==1.22.0                 # HuggingFace → ONNX export

# Sentence Transformers (relevance + topic guardrail)
sentence-transformers==3.1.1

# Detoxify (toxicity scoring)
detoxify==0.5.2

# PII Detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.6
# After pip install, run: python -m spacy download en_core_web_lg

# Prompt Injection (no extra install — uses transformers pipeline)
# Model: deepset/deberta-v3-base-injection (downloaded at first run)

# ── Cloud LLM SDKs ────────────────────────────────────────────────────────────
openai>=1.50.0               # OpenAI / OpenAI-compatible provider
anthropic>=0.37.0            # Anthropic Claude provider
google-genai>=1.0.0          # Google Gemini provider (replaces deprecated google-generativeai)

# ── Observability ──────────────────────────────────────────────────────────────
prometheus-client>=0.21.0    # /metrics Prometheus endpoint
python-json-logger>=2.0.7    # structured JSON logging for production
